+++ setup.cfg
@@ -4,14 +4,22 @@ exclude = .git,__pycache__,vendor,.venv,build
 max_line_length = 88
 
 # enforce single quotes
-select = Q0
 docstring-quotes = '''
 multiline-quotes = '''
 
-ignore = X100, E262
+extend-ignore =
+    X100, E262,
+    # See https://github.com/PyCQA/pycodestyle/issues/373
+    E203,
+    # comply with black
+    # it will handle line lengths automatically
+    E501,
+    # it also controls trailing commas in general
+    C812, C813, C815,
+
 per-file-ignores =
 # Allow crazy line lengths, unused variables, and multiple spaces after commas in lists (for grid alignment)
-	user_keymaps/**/*.py: F401,E501,E241,E131
+	user_keymaps/**/*.py: F401,E501,E241,E131,BLK100
 	tests/test_data/keymaps/**/*.py: F401,E501
 # Forgive me for my RAM hack sins
 	kmk/keyboard_config.py: I001,I003,I004,F401
